markov-writer/pyproject.toml
[tool.poetry]
name = "markov-writer"
version = "0.1.0"
description = ""
authors = ["ssilvidi <steve.silvidi@rackspace.com>"]

[tool.poetry.dependencies]
python = "^3.8"
markovify = "^0.9.3"
python-slugify = "^5.0.2"
GitPython = "^3.1.24"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
